Refactor adapters and enhance `LibreTranslateClient` reliability
- Consolidated adapter logic by removing `libretranslate_adapter.py` and moving `TranslationAdapter` protocol to `libretranslate.py`.
- Implemented `available` check in `LibreTranslateClient` to verify server status before translation.
- Upgraded `HttpClient` to support GET requests and generic method handling via `_request`.
- Updated `SimpleHttpResponse` to handle list payloads and empty response content.
- Corrected `DockerManager` import path in `translator/cli/app.py`.
